don't DELETE rows of players who have played in matches

+++ app/api/delete-player/route.ts
@@ -11,13 +11,25 @@ export async function POST(request: Request) {
         { status: 400 }
       );
     }
-    console.log("Deleting player: ", id);
-    await sql`
-      DELETE FROM players
-      WHERE id = ${id}
+
+    const result = await sql`
+      SELECT 1 FROM match_results WHERE player1 = ${id} OR player2 = ${id} LIMIT 1
     `;
 
-    return NextResponse.json({ message: "Player deleted successfully" });
+    if (result.rowCount && result.rowCount > 0) {
+      await sql`
+        UPDATE players SET is_deleted = true WHERE id = ${id}
+      `;
+      return NextResponse.json({
+        message: "Player marked as deleted because they are in a match",
+      });
+    } else {
+      await sql`
+        DELETE FROM players
+        WHERE id = ${id}
+      `;
+      return NextResponse.json({ message: "Player deleted successfully" });
+    }
   } catch (error) {
     console.error("Failed to delete player:", error);
     return NextResponse.json(
